Refactor: Car 클래스 폴더 안으로 분리
